lifting wandb integration into batteries
